+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add props `is_hoverable` and `href` to `BNavbarDropdownItem`.
+
 ## [0.0.0-alpha.13] - 2024-04-19
 
 ### Changed